Unify fixtures
